moved transferAccount reducer to follow common pattern
